requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiormq==6.2.2
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.3.4
appdirs==1.4.3
asgiref==3.5.0
aspy.refactor-imports==2.1.0
astroid==2.4.2
asttokens==2.0.5
async-timeout==4.0.2
asyncpg==0.25.0
atlastk==0.13
atomicwrites==1.4.0
attrs==19.3.0
autoflake==1.3.1
Babel==2.9.1
bandit==1.6.2
black==19.10b0
cached-property==1.5.1
certifi==2021.10.8
cfgv==3.1.0
channel-box==0.3.1
chardet==3.0.4
charset-normalizer==2.0.12
ciso8601==2.1.3
click==7.1.1
colorama==0.4.3
dead==1.3.0
distlib==0.3.0
entrypoints==0.3
eradicate==1.0
executing==0.8.1
filelock==3.0.12
flake8==3.7.9
flake8-annotations==2.1.0
flake8-annotations-coverage==0.0.3
flake8-bandit==2.1.2
flake8-broken-line==0.2.0
flake8-bugbear==20.1.4
flake8-builtins==1.5.2
flake8-commas==2.0.0
flake8-comprehensions==3.2.2
flake8-debugger==3.2.1
flake8-eradicate==0.3.0
flake8-executable==2.0.3
flake8-fixme==1.1.1
flake8-future-import==0.4.6
flake8-mutable==1.2.0
flake8-plugin-utils==1.3.0
flake8-polyfill==1.0.2
flake8-pyi==19.3.0
flake8-pytest==1.3
flake8-pytest-style==1.1.1
flake8-string-format==0.3.0
flake8-todo==0.7
flake8-unused-arguments==0.0.3
frozenlist==1.3.0
gitdb==4.0.4
GitPython==3.1.1
h11==0.9.0
httptools==0.1.2
identify==1.4.15
idna==2.10
iso8601==0.1.16
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
mailjet-rest==1.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.2.0
multidict==5.0.0
nodeenv==1.3.5
packaging==20.3
pamqp==3.1.0
passlib==1.7.2
pathspec==0.8.0
pbr==5.4.5
pluggy==0.13.1
pre-commit==2.3.0
py==1.8.1
pycodestyle==2.5.0
pydantic==1.7.2
pyflakes==2.1.1
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.7
PyPika==0.37.1
pypika-tortoise==0.1.3
pytest==5.4.2
pytz==2021.3
pyupgrade==2.3.0
PyYAML==5.3.1
regex==2020.4.4
reorder-python-imports==2.2.0
requests==2.24.0
simple-print==1.5
six==1.14.0
smmap==3.0.2
sniffio==1.2.0
starlette==0.16.0
stevedore==1.32.0
termcolor==1.1.0
tokenize-rt==4.0.0
toml==0.10.0
tortoise==0.1.1
tortoise-orm==0.18.1
typed-ast==1.5.2
typing-extensions==3.7.4.2
ua-parser==0.10.0
urllib3==1.25.11
user-agents==2.2.0
uvicorn==0.11.3
uvloop==0.16.0
virtualenv==20.0.18
wcwidth==0.1.9
websockets==8.1
wrapt==1.12.1
yarl==1.6.2
yesqa==1.1.0
